feature: implement yaml.Marshaller on Tag
Only really necessary to do custom marshal/unmarshal to avoid storing
the normalised forms of all the Parents, given it is a generic map. Were
it a custom type, or maybe a set from a 3rd party, not really necessary
rn. Not even that vital to not store them, but saves a little on-disk
space at least